\name{latlon2idx}
\alias{latlon2idx}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ latlon2idx}
\description{
  A sequential index number based on lat/lon and a reference grid.
}
\usage{
latlon2idx(lat, lon, layer)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lat}{ Latitude }
  \item{lon}{ Longitude }
  \item{layer}{ A reference layer. }
}

\value{
 Returns a an index number.
}
\references{ none }
\author{ Reinhard Simon }
\keyword{ manip }% __ONLY ONE__ keyword per line
